refactor: Payment 에서 Double 로 형변환하지 않고 사용하는 곳에서 형변환해서 사용하도록 수정

+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/domain/payment/Payment.kt
@@ -9,5 +9,5 @@ class Payment(
 
     fun getPaidTicketCount() = paidMoney / PRICE
 
-    fun getPaidMoney() = paidMoney.toDouble()
+    fun getPaidMoney() = paidMoney
 }
